requirements.txt
cffi==1.17.1
cryptography==44.0.2
greenlet==3.1.1
pillow==11.2.1
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
ttkbootstrap==1.12.0
typing_extensions==4.13.2
